[bundyfork] fix test that relies on ordering of internal data structure.
We seem to be relying on the order of data returned, which is not always
guaranteed. My guess is that this is some sort of hash, and that the
most recent version of Python returns these in more random order to
prevent dictionary collision attacks (different each run of a program).