[shmemfix3] update auth so it ignores segment update with inuse-only when not.
This can happen in normal initialization, i.e., both the authoritative
server and the memory manager just starts.  This initial reset command
is just for segment reader modules to be synchronized with the memory
manager, and the actual reset will be delayed until the memory manager
completes building the segment data with any un-applied updates.