[shmemfix3] updated mapped zt segment to allow clearing the segment.
this will be needed if memmgr fails to initialize the segment data on
restart, while reader modules are currently somehow using the segment.
in this case memmgr will send a segment info update to reset the
segment for synchronization, and, as a result, readers will fail to use
the segment until it's fixed.  that's not good, but this should be a rare
erroneous condition and should be acceptable.