[shmem4] add param to complete_update to indicate success/failure of update.
and if the update fails in the UPDATING state, move back to READY, instead
of trying to synchronize with readers and COPY the data.  in case diff-based
update fails at the last stage (which invalidates the zone followed by full
reload), it would be better to let the readers keep using the current version
until the next load happens.