[shmemfix7] introduced a new "SYNCHRONIZING2" state for segment info.
for the initial load, it's probably better to make sure that readers
start using after both read and write versions have been fully loaded.
otherwise, it could be possible that loading of the writer version may take
long while new updates are coming.

also, changed the behavior on transitioning to READY; if an outstanding
event exists at that point, immediately start it.  otherwise this event
will be pending until next event happens, which is not good.