[master] make sure xfrout_server inited in main() is handled as global.
otherwise resetting it to None in shutdown() doesn't work as intended
and results in duplicate shutdown().  this seems to be a regression
when this block of code was moved to a separate function (main()).

(more fundamentally, we should rather stop relying on such a wider scope
variable.  but I'd leave it to a next step).

should be trivial enough, committing it directly to master.