[xfrout] (big) refactoring: use BUNDYServer to eliminate intermediate thread
It also uses the generic socketsession module to get xfr queries from
bundy-auth.  this will eliminate the need for the in-house and yet incomplete
support of query forwarding.

a new tentative class NewXfroutServer is introduced to add the implementation
and unit tests until the auth side is updated (it's currently only used for
tests).  This is mostly a copy + adjustment of the original xfrout and
UnixSockServer classes, and of bundy-memmgr (which uses BUNDYServer).